Generate Areas after login
